Add route, display single game card.
